#5 - Refactored the code and orgnized the folders.
